Add library for string-trimming
The use of this function helps remove unnecessary spaces and newline
characters from user-inputted strings.
At the same time, I have also modified the previous code to incorporate the
use of this function.
