+++ common/inc/string_utils.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void trim_string(char *str);  // trimming off space and newline in string
