+++ server/handler.c
@@ -6,6 +6,7 @@
 #include "history.h"
 #include "login.h"
 #include "proto_lib.h"
+#include "string_utils.h"
 
 #define UPDATE_PERIOD 100000
 static uint32_t UUID = 0;
@@ -458,8 +459,8 @@ const func func_table[] = {
 
 void server_select_cmd(int connfd, char *sendline, void *hdl_obj)
 {
-    char *token, *dummy = sendline;
-    token = strtok(dummy, " \n");
+    char *token = sendline;
+    trim_string(token);
     if (token != NULL) {
         uint8_t cmd_index = 0;
         int ret = sscanf(token, "%hhu", &cmd_index);
