+++ server/login.c
@@ -1,38 +1,12 @@
 #include "login.h"
 #include "error.h"
+#include "string_utils.h"
 
 User user_database[MAX_USERS] = {
     {"root", "root"},
     {"computer", "network"},
     // ... add it to 10 :)
 };
-
-// Remove leading and trailing spaces and newline characters from a string
-void trim_string(char *str)
-{
-    size_t len = strlen(str);
-
-    // Remove leading spaces and newline characters
-    size_t start = 0;
-    while (start < len && (str[start] == ' ' || str[start] == '\n')) {
-        ++start;
-    }
-
-    // Remove trailing spaces and newline characters
-    size_t end = len - 1;
-    while (end > start && (str[end] == ' ' || str[end] == '\n')) {
-        --end;
-    }
-
-    // Move non-empty characters to the beginning of the string
-    size_t i;
-    for (i = start; i <= end; ++i) {
-        str[i - start] = str[i];
-    }
-
-    // Add the string terminator
-    str[i - start] = '\0';
-}
 
 // Authenticate a user by comparing the provided username and password
 bool authenticate_user(const char *username, const char *password)
