+++ server/login.h
@@ -12,4 +12,3 @@ typedef struct {
 } User;
 
 bool authenticate_user(const char *username, const char *password);
-void trim_string(char *str);
